Add Alpaca outcome tracking, closing the last deferred gap
Builds the piece explicitly flagged as deferred when the Alpaca
integration first shipped: closed Alpaca trades now get linked back to
the trade_intent that caused them, same as OANDA's tracker does, so
Alpaca trades will show up in Trade History and feed the track-record
system once they accumulate.

Alpaca has no OANDA-style transaction ledger with an explicit
tradesClosed[] link, so this works differently: it walks this system's
own orders_fills rows (every entry placed via ExecutionService, with the
real broker_order_id/client_order_id Alpaca gave it) and asks Alpaca
directly whether each has filled and resolved an exit yet - a bracket
leg for equities (Alpaca's own OCO), or the separate stop_limit order
this system attaches after a crypto fill (src/broker/alpaca.py).

Real gap found live: Alpaca's bulk closed-orders list flattens a
bracket's legs into separate top-level entries instead of nesting them,
so it can't answer "did this specific order's exit fill" - added
AlpacaBroker.get_order() (single-order fetch, which does nest legs) and
used that instead for anything that needs one order's real state.

Also fixed, now that Alpaca outcomes are real instead of hypothetical:
- trade_outcomes.units was still Integer from before this column existed
  for Alpaca - a real fractional crypto close (0.00029925 BTC) silently
  truncated to 0 in testing; widened to Float and fixed live.
- train_meta_model.py and promotion_gates.py both read ALL trade_outcomes
  with no broker filter - once Alpaca trades really do close, they'd
  silently pool equity/crypto outcomes into the shared FX meta-model's
  training data (whose macro/news/cross-market/session components are all
  trivially zero for non-forex, see this session's earlier guards) and
  into what is explicitly documented as the OANDA-specific promotion
  report. Both filtered to broker='oanda' until there's a real reason for
  a joint or Alpaca-specific model.
- run_trade_attribution.py hardcoded OandaBroker for every outcome's
  regime-change check - now routes per instrument like every other
  broker-dispatch fix this session.

Verified live end-to-end: a real crypto entry+stop-loss round trip
(through the actual AlpacaBroker.place_order() path) correctly produced
a trade_outcomes row via sync_alpaca_outcomes with the right entry/exit
price, fractional units, and P&L; the real scheduled sync_outcomes.py and
run_trade_attribution.py scripts both ran clean against the live account.

+++ src/data/db.py
@@ -429,7 +429,13 @@ trade_outcomes = Table(
     Column("execution_mode", String, nullable=False),  # paper / demo
     Column("instrument", String, nullable=False),
     Column("action", String, nullable=False),  # BUY / SELL
-    Column("units", Integer, nullable=False),
+    # Float, not Integer: crypto (Alpaca) closes in fractional units — real
+    # bug found live 2026-08-22, a genuine 0.00029925 BTC close silently
+    # truncated to 0 in this column while orders_fills.units/risk_decisions.
+    # size_units had already been widened for the same reason (Phase 6 of
+    # the Alpaca integration missed this one since Alpaca outcome tracking
+    # didn't exist yet at that point).
+    Column("units", Float, nullable=False),
     # nullable: real production gap found live 2026-08-18 — OANDA's
     # transaction ledger doesn't always retain the ORDER_FILL that opened a
     # trade long enough for src/outcomes/tracker.py's sync to find it
